webapp/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
openai==1.3.5
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
moviepy==1.0.3
googletrans==4.0.0rc1
gTTS==2.4.0
pydub==0.25.1
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.1.0
requests==2.31.0

